backend/requirements.txt
tensorflow
opencv-python
numpy
scikit-learn
matplotlib
Flask
Flask-Cors
Pillow
librosa